Add 16-bit checksum (instead of 8-bit checksum)
8-bit checksum lets a lot of garbage through!

Discovered when using 900 MHz radios.

+++ include/simple_serial/SimpleMavlink.h
@@ -9,7 +9,7 @@
     uint8_t magic; \
     uint8_t msg_id;
 #define MSG_END \
-    uint8_t csum; \
+    uint16_t csum; \
   };
 
 MSG_START(rpm, 0)
